pyproject.toml
[tool.poetry]
authors = ["shoor <shoorday@gmail.com>"]
description = ""
license = "MIT"
name = "xdutools"
version = "0.1.0"

[tool.poetry.dependencies]
beautifulsoup4 = "^4.9.3"
click = "^7.1.0"
httpx = "^0.16.1"
lxml = {version = "^4.6.1", optional = true}
pycryptodome = "^3.9.9"
python = "^3.9"
requests = "^2.25"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
mypy = "^0.790"

[tool.poetry.extras]
all = ["lxml"]

[[tool.poetry.source]]
default = true
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"

[tool.poetry.scripts]
xdu = "xdutools.cli:main"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.black]
line-length = 88
